requirements.txt
django~=3.1
django-annoying~=0.10
django-pipeline~=2.0
psycopg2-binary~=2.8
requests~=2.24
uwsgi~=2.0
httpretty~=1.0
# tests
model_mommy~=2.0
django-environ~=0.4
# api
djangorestframework<3.12
drf-yasg~=1.17 # docs generation
packaging~=20.4
